requirements.txt
djangorestframework-simplejwt=5.3.1
django-cros-headers=4.3.1
djangorestframework=3.15.1